USDLoader: Refactored and added USDComposer class.
