+++ examples/jsm/loaders/usd/USDCParser.js
@@ -1,26 +1,3 @@
-import {
-	AnimationClip,
-	Bone,
-	BufferAttribute,
-	BufferGeometry,
-	ClampToEdgeWrapping,
-	Euler,
-	Group,
-	Matrix4,
-	NoColorSpace,
-	Mesh,
-	MeshPhysicalMaterial,
-	MirroredRepeatWrapping,
-	QuaternionKeyframeTrack,
-	RepeatWrapping,
-	Skeleton,
-	SkinnedMesh,
-	SRGBColorSpace,
-	TextureLoader,
-	Object3D,
-	VectorKeyframeTrack
-} from 'three';
-
 const textDecoder = new TextDecoder();
 
 // Type enum values from crateDataTypes.h
@@ -538,15 +515,15 @@ class ValueRep {
 
 class USDCParser {
 
-	parse( buffer, assets = {}, variantSelections = {} ) {
+	/**
+	 * Parse USDC file and return raw spec data without building Three.js scene.
+	 * Used by USDComposer for unified scene composition.
+	 */
+	parseData( buffer ) {
 
 		this.buffer = buffer instanceof ArrayBuffer ? buffer : buffer.buffer;
 		this.reader = new BinaryReader( this.buffer );
-		this.assets = assets;
-		this.externalVariantSelections = variantSelections;
 		this.version = { major: 0, minor: 0, patch: 0 };
-		this.textureLoader = new TextureLoader();
-		this.textureCache = {};
 
 		this._readBootstrap();
 		this._readTOC();
@@ -557,7 +534,20 @@ class USDCParser {
 		this._readPaths();
 		this._readSpecs();
 
-		return this._buildScene();
+		// Build specsByPath without building scene
+		this.specsByPath = {};
+
+		for ( const spec of this.specs ) {
+
+			const path = this.paths[ spec.pathIndex ];
+			if ( ! path ) continue;
+
+			const fields = this._getFieldsForSpec( spec );
+			this.specsByPath[ path ] = { specType: spec.specType, fields };
+
+		}
+
+		return { specsByPath: this.specsByPath };
 
 	}
 
@@ -1172,6 +1162,38 @@ class USDCParser {
 			case TypeEnum.Permission:
 			case TypeEnum.Variability:
 				return payload;
+
+			// Inlined vectors are encoded as signed 8-bit integers
+			case TypeEnum.Vec2f:
+			case TypeEnum.Vec2i: {
+
+				const buf = new ArrayBuffer( 4 );
+				const view = new DataView( buf );
+				view.setUint32( 0, payload, true );
+				return [ view.getInt8( 0 ), view.getInt8( 1 ) ];
+
+			}
+
+			case TypeEnum.Vec3f:
+			case TypeEnum.Vec3i: {
+
+				const buf = new ArrayBuffer( 4 );
+				const view = new DataView( buf );
+				view.setUint32( 0, payload, true );
+				return [ view.getInt8( 0 ), view.getInt8( 1 ), view.getInt8( 2 ) ];
+
+			}
+
+			case TypeEnum.Vec4f:
+			case TypeEnum.Vec4i: {
+
+				const buf = new ArrayBuffer( 4 );
+				const view = new DataView( buf );
+				view.setUint32( 0, payload, true );
+				return [ view.getInt8( 0 ), view.getInt8( 1 ), view.getInt8( 2 ), view.getInt8( 3 ) ];
+
+			}
+
 			default:
 				return payload;
 
@@ -1674,41 +1696,6 @@ class USDCParser {
 
 	}
 
-	// ========================================================================
-	// Scene Building
-	// ========================================================================
-
-	_buildScene() {
-
-		this.specsByPath = {};
-
-		for ( const spec of this.specs ) {
-
-			const path = this.paths[ spec.pathIndex ];
-			if ( ! path ) continue;
-
-			const fields = this._getFieldsForSpec( spec );
-			this.specsByPath[ path ] = { specType: spec.specType, fields };
-
-		}
-
-		const rootSpec = this.specsByPath[ '/' ];
-		const rootFields = rootSpec ? rootSpec.fields : {};
-		this.fps = rootFields.framesPerSecond || rootFields.timeCodesPerSecond || 30;
-
-		this.skeletons = {};
-		this.skinnedMeshes = [];
-
-		const group = new Group();
-		this._buildHierarchy( group, '/' );
-		this._bindSkeletons();
-
-		group.animations = this._buildAnimations();
-
-		return group;
-
-	}
-
 	_getFieldsForSpec( spec ) {
 
 		const fields = {};
@@ -1741,1841 +1728,6 @@ class USDCParser {
 		}
 
 		return fields;
-
-	}
-
-	_buildHierarchy( parent, parentPath ) {
-
-		const prefix = parentPath === '/' ? '/' : parentPath + '/';
-
-		// Build list of variant paths to also search for children
-		const parentSpec = this.specsByPath[ parentPath ];
-		const variantSetChildren = parentSpec && parentSpec.fields ? parentSpec.fields.variantSetChildren : null;
-		const variantPaths = [];
-
-		if ( variantSetChildren && variantSetChildren.length > 0 ) {
-
-			for ( const variantSetName of variantSetChildren ) {
-
-				// External variant selections (from referencing prim) take priority
-				let selectedVariant = this.externalVariantSelections[ variantSetName ] || null;
-
-				// Fall back to file's internal variant selection
-				if ( ! selectedVariant ) {
-
-					const variantSelection = parentSpec.fields.variantSelection;
-					selectedVariant = variantSelection ? variantSelection[ variantSetName ] : null;
-
-				}
-
-				// Fall back to first variant child
-				if ( ! selectedVariant ) {
-
-					const variantSetPath = parentPath + '/{' + variantSetName + '=}';
-					const variantSetSpec = this.specsByPath[ variantSetPath ];
-					if ( variantSetSpec && variantSetSpec.fields && variantSetSpec.fields.variantChildren ) {
-
-						selectedVariant = variantSetSpec.fields.variantChildren[ 0 ];
-
-					}
-
-				}
-
-				if ( selectedVariant ) {
-
-					const variantPath = parentPath + '/{' + variantSetName + '=' + selectedVariant + '}';
-					variantPaths.push( variantPath );
-
-				}
-
-			}
-
-		}
-
-		for ( const path in this.specsByPath ) {
-
-			const spec = this.specsByPath[ path ];
-
-			let isChild = this._isDirectChild( parentPath, path, prefix );
-			if ( ! isChild ) {
-
-				for ( const vp of variantPaths ) {
-
-					const vpPrefix = vp + '/';
-					if ( this._isDirectChild( vp, path, vpPrefix ) ) {
-
-						isChild = true;
-						break;
-
-					}
-
-				}
-
-			}
-
-			if ( ! isChild ) continue;
-
-			// Only process Prim specs
-			if ( spec.specType !== SpecType.Prim ) continue;
-
-			const specifier = spec.fields.specifier;
-			if ( specifier !== Specifier.Def ) continue;
-
-			const typeName = spec.fields.typeName || '';
-			const name = this._getPathName( path );
-
-			if ( typeName === 'SkelRoot' ) {
-
-				// Skeletal root - treat as transform but track for skeleton binding
-				const obj = this._buildXform( path, spec );
-				obj.name = name;
-				obj.userData.isSkelRoot = true;
-				parent.add( obj );
-
-				// Recursively build children
-				this._buildHierarchy( obj, path );
-
-			} else if ( typeName === 'Skeleton' ) {
-
-				// Build skeleton and store it
-				const skeleton = this._buildSkeleton( path );
-				if ( skeleton ) {
-
-					this.skeletons[ path ] = skeleton;
-
-				}
-
-				// Recursively build children (may contain SkelAnimation)
-				this._buildHierarchy( parent, path );
-
-			} else if ( typeName === 'SkelAnimation' ) {
-
-				// Skip - animations are processed separately in _buildAnimations
-
-			} else if ( typeName === 'Xform' || typeName === 'Scope' || typeName === '' ) {
-
-				// Transform node or group
-				const obj = this._buildXform( path, spec );
-				obj.name = name;
-				parent.add( obj );
-
-				// Recursively build children
-				this._buildHierarchy( obj, path );
-
-			} else if ( typeName === 'Mesh' ) {
-
-				// Mesh (may be skinned)
-				const mesh = this._buildMesh( path, spec );
-				mesh.name = name;
-				parent.add( mesh );
-
-			} else if ( typeName === 'Material' || typeName === 'Shader' ) {
-
-				// Skip materials/shaders, they're referenced by meshes
-
-			} else {
-
-				// Unknown type, create empty object and recurse
-				const obj = new Object3D();
-				obj.name = name;
-				parent.add( obj );
-				this._buildHierarchy( obj, path );
-
-			}
-
-		}
-
-	}
-
-	_isDirectChild( parentPath, childPath, prefix ) {
-
-		if ( parentPath === '/' ) {
-
-			// Root children: /Name (no additional slashes)
-			return childPath.startsWith( '/' ) &&
-				childPath.indexOf( '/', 1 ) === - 1 &&
-				childPath.length > 1;
-
-		}
-
-		// Must start with parent path
-		if ( ! childPath.startsWith( prefix ) ) return false;
-
-		// Must not have additional slashes (direct child only)
-		const remainder = childPath.slice( prefix.length );
-		return remainder.indexOf( '/' ) === - 1 && remainder.length > 0;
-
-	}
-
-	_getPathName( path ) {
-
-		const lastSlash = path.lastIndexOf( '/' );
-		return lastSlash >= 0 ? path.slice( lastSlash + 1 ) : path;
-
-	}
-
-	_buildXform( path, spec ) {
-
-		const obj = new Object3D();
-
-		// Get attribute values from child attribute specs (for transforms)
-		const attrs = this._getAttributeValues( path );
-
-		// Apply transform
-		this._applyTransform( obj, spec.fields, attrs );
-
-		return obj;
-
-	}
-
-	_buildMesh( path, spec ) {
-
-		// Get attribute values from child attribute specs
-		const attrs = this._getAttributeValues( path );
-
-		// Check for skinning data
-		const jointIndices = attrs[ 'primvars:skel:jointIndices' ];
-		const jointWeights = attrs[ 'primvars:skel:jointWeights' ];
-		const hasSkinning = jointIndices && jointWeights &&
-			jointIndices.length > 0 && jointWeights.length > 0;
-
-		// Collect GeomSubsets for multi-material support
-		const geomSubsets = this._getGeomSubsets( path );
-
-		let geometry, material;
-
-		if ( geomSubsets.length > 0 ) {
-
-			// Multi-material mesh: reorder triangles by material group
-			geometry = this._buildGeometryWithSubsets( attrs, geomSubsets, hasSkinning );
-			material = geomSubsets.map( subset => this._buildMaterialForPath( subset.materialPath ) );
-
-		} else {
-
-			// Single material mesh
-			geometry = this._buildGeometry( path, attrs, hasSkinning );
-			material = this._buildMaterial( path, spec.fields );
-
-		}
-
-		// Apply displayColor if no texture/color was set
-		const displayColor = attrs[ 'primvars:displayColor' ];
-		if ( displayColor && displayColor.length >= 3 ) {
-
-			const applyDisplayColor = ( mat ) => {
-
-				if ( mat.color && mat.color.r === 1 && mat.color.g === 1 && mat.color.b === 1 && ! mat.map ) {
-
-					mat.color.setRGB( displayColor[ 0 ], displayColor[ 1 ], displayColor[ 2 ] );
-
-				}
-
-			};
-
-			if ( Array.isArray( material ) ) {
-
-				material.forEach( applyDisplayColor );
-
-			} else {
-
-				applyDisplayColor( material );
-
-			}
-
-		}
-
-		let mesh;
-
-		if ( hasSkinning ) {
-
-			mesh = new SkinnedMesh( geometry, material );
-
-			// Find skeleton path from skel:skeleton relationship
-			const skelBindingPath = path + '.skel:skeleton';
-			const skelBindingSpec = this.specsByPath[ skelBindingPath ];
-			let skeletonPath = null;
-
-			if ( skelBindingSpec && skelBindingSpec.fields.targetPaths && skelBindingSpec.fields.targetPaths.length > 0 ) {
-
-				skeletonPath = skelBindingSpec.fields.targetPaths[ 0 ];
-
-			}
-
-			// Get per-mesh joint mapping (local joint names for this mesh)
-			const localJoints = attrs[ 'skel:joints' ];
-
-			// Track for later skeleton binding
-			this.skinnedMeshes.push( { mesh, skeletonPath, path, localJoints } );
-
-		} else {
-
-			mesh = new Mesh( geometry, material );
-
-		}
-
-		// Apply transforms to the Object3D (handles both simple and pivot transforms)
-		this._applyTransform( mesh, spec.fields, attrs );
-
-		return mesh;
-
-	}
-
-	_getGeomSubsets( meshPath ) {
-
-		const subsets = [];
-		const prefix = meshPath + '/';
-
-		for ( const p in this.specsByPath ) {
-
-			if ( ! p.startsWith( prefix ) ) continue;
-
-			const spec = this.specsByPath[ p ];
-			if ( spec.fields.typeName !== 'GeomSubset' ) continue;
-
-			const attrs = this._getAttributeValues( p );
-			const indices = attrs[ 'indices' ];
-			if ( ! indices || indices.length === 0 ) continue;
-
-			// Get material binding
-			const bindingPath = p + '.material:binding';
-			const bindingSpec = this.specsByPath[ bindingPath ];
-			let materialPath = null;
-			if ( bindingSpec && bindingSpec.fields.targetPaths && bindingSpec.fields.targetPaths.length > 0 ) {
-
-				materialPath = bindingSpec.fields.targetPaths[ 0 ];
-
-			}
-
-			subsets.push( {
-				name: this._getPathName( p ),
-				indices: indices, // face indices
-				materialPath: materialPath
-			} );
-
-		}
-
-		return subsets;
-
-	}
-
-	_buildGeometryWithSubsets( fields, geomSubsets, hasSkinning = false ) {
-
-		const geometry = new BufferGeometry();
-
-		const points = fields[ 'points' ];
-		if ( ! points || points.length === 0 ) return geometry;
-
-		const faceVertexIndices = fields[ 'faceVertexIndices' ];
-		const faceVertexCounts = fields[ 'faceVertexCounts' ];
-
-		if ( ! faceVertexCounts || faceVertexCounts.length === 0 ) return geometry;
-
-		const { uvs, uvIndices } = this._findUVPrimvar( fields );
-		const normals = fields[ 'normals' ] || fields[ 'primvars:normals' ];
-
-		const jointIndices = hasSkinning ? fields[ 'primvars:skel:jointIndices' ] : null;
-		const jointWeights = hasSkinning ? fields[ 'primvars:skel:jointWeights' ] : null;
-		const elementSize = fields[ 'primvars:skel:jointIndices:elementSize' ] || 4;
-
-		// Build face-to-triangle mapping (triangles per face and cumulative offset)
-		const faceTriangleOffset = [];
-		let triangleCount = 0;
-
-		for ( let i = 0; i < faceVertexCounts.length; i ++ ) {
-
-			faceTriangleOffset.push( triangleCount );
-			const count = faceVertexCounts[ i ];
-			if ( count >= 3 ) triangleCount += count - 2;
-
-		}
-
-		const triangleToSubset = new Int32Array( triangleCount ).fill( - 1 );
-
-		for ( let si = 0; si < geomSubsets.length; si ++ ) {
-
-			const subset = geomSubsets[ si ];
-
-			for ( let i = 0; i < subset.indices.length; i ++ ) {
-
-				const faceIdx = subset.indices[ i ];
-				if ( faceIdx >= faceVertexCounts.length ) continue;
-
-				const triStart = faceTriangleOffset[ faceIdx ];
-				const triCount = faceVertexCounts[ faceIdx ] - 2;
-
-				for ( let t = 0; t < triCount; t ++ ) {
-
-					triangleToSubset[ triStart + t ] = si;
-
-				}
-
-			}
-
-		}
-
-		// Sort triangles by subset (unassigned first, then by subset index)
-		const sortedTriangles = [];
-
-		for ( let tri = 0; tri < triangleCount; tri ++ ) {
-
-			sortedTriangles.push( { original: tri, subset: triangleToSubset[ tri ] } );
-
-		}
-
-		sortedTriangles.sort( ( a, b ) => a.subset - b.subset );
-		const groups = [];
-		let currentSubset = sortedTriangles.length > 0 ? sortedTriangles[ 0 ].subset : - 1;
-		let groupStart = 0;
-
-		for ( let i = 0; i < sortedTriangles.length; i ++ ) {
-
-			if ( sortedTriangles[ i ].subset !== currentSubset ) {
-
-				if ( currentSubset >= 0 ) {
-
-					groups.push( {
-						start: groupStart * 3,
-						count: ( i - groupStart ) * 3,
-						materialIndex: currentSubset
-					} );
-
-				}
-
-				currentSubset = sortedTriangles[ i ].subset;
-				groupStart = i;
-
-			}
-
-		}
-
-		// Add final group
-		if ( currentSubset >= 0 && sortedTriangles.length > groupStart ) {
-
-			groups.push( {
-				start: groupStart * 3,
-				count: ( sortedTriangles.length - groupStart ) * 3,
-				materialIndex: currentSubset
-			} );
-
-		}
-
-		// Apply groups to geometry
-		for ( const group of groups ) {
-
-			geometry.addGroup( group.start, group.count, group.materialIndex );
-
-		}
-
-		// Triangulate original data
-		const origIndices = this._triangulateIndices( faceVertexIndices, faceVertexCounts );
-		const origUvIndices = uvIndices ? this._triangulateIndices( uvIndices, faceVertexCounts ) : null;
-
-		// Triangulate normals if they are faceVarying (one per face-vertex)
-		const numFaceVertices = faceVertexCounts.reduce( ( a, b ) => a + b, 0 );
-		const hasFaceVaryingNormals = normals && normals.length / 3 === numFaceVertices;
-		const origNormalIndices = hasFaceVaryingNormals
-			? this._triangulateIndices( Array.from( { length: numFaceVertices }, ( _, i ) => i ), faceVertexCounts )
-			: null;
-
-		// Build reordered vertex data
-		const vertexCount = triangleCount * 3;
-		const positions = new Float32Array( vertexCount * 3 );
-		const uvData = uvs ? new Float32Array( vertexCount * 2 ) : null;
-		const normalData = normals ? new Float32Array( vertexCount * 3 ) : null;
-		const skinIndexData = jointIndices ? new Uint16Array( vertexCount * 4 ) : null;
-		const skinWeightData = jointWeights ? new Float32Array( vertexCount * 4 ) : null;
-
-		for ( let i = 0; i < sortedTriangles.length; i ++ ) {
-
-			const origTri = sortedTriangles[ i ].original;
-
-			for ( let v = 0; v < 3; v ++ ) {
-
-				const origIdx = origTri * 3 + v;
-				const newIdx = i * 3 + v;
-
-				// Position
-				const pointIdx = origIndices[ origIdx ];
-				positions[ newIdx * 3 ] = points[ pointIdx * 3 ];
-				positions[ newIdx * 3 + 1 ] = points[ pointIdx * 3 + 1 ];
-				positions[ newIdx * 3 + 2 ] = points[ pointIdx * 3 + 2 ];
-
-				// UVs
-				if ( uvData && uvs ) {
-
-					if ( origUvIndices ) {
-
-						const uvIdx = origUvIndices[ origIdx ];
-						uvData[ newIdx * 2 ] = uvs[ uvIdx * 2 ];
-						uvData[ newIdx * 2 + 1 ] = uvs[ uvIdx * 2 + 1 ];
-
-					} else if ( uvs.length / 2 === points.length / 3 ) {
-
-						// Per-vertex UVs
-						uvData[ newIdx * 2 ] = uvs[ pointIdx * 2 ];
-						uvData[ newIdx * 2 + 1 ] = uvs[ pointIdx * 2 + 1 ];
-
-					}
-
-				}
-
-				// Normals
-				if ( normalData && normals ) {
-
-					if ( origNormalIndices ) {
-
-						// FaceVarying normals
-						const normalIdx = origNormalIndices[ origIdx ];
-						normalData[ newIdx * 3 ] = normals[ normalIdx * 3 ];
-						normalData[ newIdx * 3 + 1 ] = normals[ normalIdx * 3 + 1 ];
-						normalData[ newIdx * 3 + 2 ] = normals[ normalIdx * 3 + 2 ];
-
-					} else if ( normals.length === points.length ) {
-
-						// Per-vertex normals
-						normalData[ newIdx * 3 ] = normals[ pointIdx * 3 ];
-						normalData[ newIdx * 3 + 1 ] = normals[ pointIdx * 3 + 1 ];
-						normalData[ newIdx * 3 + 2 ] = normals[ pointIdx * 3 + 2 ];
-
-					}
-
-				}
-
-				// Skinning data
-				if ( skinIndexData && skinWeightData && jointIndices && jointWeights ) {
-
-					for ( let j = 0; j < 4; j ++ ) {
-
-						if ( j < elementSize ) {
-
-							skinIndexData[ newIdx * 4 + j ] = jointIndices[ pointIdx * elementSize + j ] || 0;
-							skinWeightData[ newIdx * 4 + j ] = jointWeights[ pointIdx * elementSize + j ] || 0;
-
-						} else {
-
-							skinIndexData[ newIdx * 4 + j ] = 0;
-							skinWeightData[ newIdx * 4 + j ] = 0;
-
-						}
-
-					}
-
-				}
-
-			}
-
-		}
-
-		geometry.setAttribute( 'position', new BufferAttribute( positions, 3 ) );
-
-		if ( uvData ) {
-
-			geometry.setAttribute( 'uv', new BufferAttribute( uvData, 2 ) );
-
-		}
-
-		if ( normalData ) {
-
-			geometry.setAttribute( 'normal', new BufferAttribute( normalData, 3 ) );
-
-		} else {
-
-			geometry.computeVertexNormals();
-
-		}
-
-		if ( skinIndexData ) {
-
-			geometry.setAttribute( 'skinIndex', new BufferAttribute( skinIndexData, 4 ) );
-
-		}
-
-		if ( skinWeightData ) {
-
-			geometry.setAttribute( 'skinWeight', new BufferAttribute( skinWeightData, 4 ) );
-
-		}
-
-		return geometry;
-
-	}
-
-	_buildMaterialForPath( materialPath ) {
-
-		const material = new MeshPhysicalMaterial();
-
-		if ( materialPath ) {
-
-			this._applyMaterial( material, materialPath );
-
-		}
-
-		return material;
-
-	}
-
-	_getAttributeValues( primPath ) {
-
-		// In USDC, attributes are stored as child specs with paths like /Mesh.points
-		// The attribute value is in the 'default' field of the attribute spec
-		const attrs = {};
-		const prefix = primPath + '.';
-
-		for ( const path in this.specsByPath ) {
-
-			// Check if this is an attribute of the prim (path contains a dot after primPath)
-			if ( ! path.startsWith( prefix ) ) continue;
-
-			const spec = this.specsByPath[ path ];
-
-			// Only process Attribute specs
-			if ( spec.specType !== SpecType.Attribute ) continue;
-
-			// Get attribute name (part after the dot)
-			const attrName = path.slice( prefix.length );
-
-			// Get the value from 'default' field
-			if ( spec.fields.default !== undefined ) {
-
-				attrs[ attrName ] = spec.fields.default;
-
-			}
-
-			// Also include elementSize for skinning attributes
-			if ( spec.fields.elementSize !== undefined ) {
-
-				attrs[ attrName + ':elementSize' ] = spec.fields.elementSize;
-
-			}
-
-			if ( attrName.startsWith( 'primvars:' ) && spec.fields.typeName !== undefined ) {
-
-				attrs[ attrName + ':typeName' ] = spec.fields.typeName;
-
-			}
-
-		}
-
-		return attrs;
-
-	}
-
-	_findUVPrimvar( fields ) {
-
-		for ( const key in fields ) {
-
-			if ( ! key.startsWith( 'primvars:' ) ) continue;
-			if ( key.endsWith( ':typeName' ) || key.endsWith( ':elementSize' ) || key.endsWith( ':indices' ) ) continue;
-			if ( key.includes( 'skel:' ) ) continue;
-
-			const typeName = fields[ key + ':typeName' ];
-			if ( typeName && typeName.includes( 'texCoord' ) ) {
-
-				return {
-					uvs: fields[ key ],
-					uvIndices: fields[ key + ':indices' ]
-				};
-
-			}
-
-		}
-
-		const uvs = fields[ 'primvars:st' ] || fields[ 'primvars:UVMap' ];
-		const uvIndices = fields[ 'primvars:st:indices' ];
-		return { uvs, uvIndices };
-
-	}
-
-	_applyTransform( obj, fields, attrs = {} ) {
-
-		// Merge fields and attrs (attrs take precedence for transforms)
-		const data = { ...fields, ...attrs };
-
-		const xformOpOrder = data[ 'xformOpOrder' ];
-
-		// If we have xformOpOrder, apply transforms in order using matrices
-		// USD uses row-vector convention (p' = p * M), Three.js uses column-vector (p' = M * p)
-		// To get equivalent transformation, iterate xformOpOrder FORWARD with right-multiply
-		if ( xformOpOrder && xformOpOrder.length > 0 ) {
-
-			const matrix = new Matrix4();
-			const tempMatrix = new Matrix4();
-
-			// Iterate FORWARD for Three.js column-vector convention
-			for ( let i = 0; i < xformOpOrder.length; i ++ ) {
-
-				const op = xformOpOrder[ i ];
-				const isInverse = op.startsWith( '!invert!' );
-				const opName = isInverse ? op.slice( 8 ) : op;
-
-				if ( opName === 'xformOp:transform' ) {
-
-					const m = data[ 'xformOp:transform' ];
-					if ( m && m.length === 16 ) {
-
-						tempMatrix.fromArray( m );
-						if ( isInverse ) tempMatrix.invert();
-						matrix.multiply( tempMatrix );
-
-					}
-
-				} else if ( opName === 'xformOp:translate' ) {
-
-					const t = data[ 'xformOp:translate' ];
-					if ( t ) {
-
-						tempMatrix.makeTranslation( t[ 0 ], t[ 1 ], t[ 2 ] );
-						if ( isInverse ) tempMatrix.invert();
-						matrix.multiply( tempMatrix );
-
-					}
-
-				} else if ( opName === 'xformOp:translate:pivot' ) {
-
-					const t = data[ 'xformOp:translate:pivot' ];
-					if ( t ) {
-
-						tempMatrix.makeTranslation( t[ 0 ], t[ 1 ], t[ 2 ] );
-						if ( isInverse ) tempMatrix.invert();
-						matrix.multiply( tempMatrix );
-
-					}
-
-				} else if ( opName === 'xformOp:scale' ) {
-
-					const s = data[ 'xformOp:scale' ];
-					if ( s ) {
-
-						if ( Array.isArray( s ) ) {
-
-							tempMatrix.makeScale( s[ 0 ], s[ 1 ], s[ 2 ] );
-
-						} else {
-
-							tempMatrix.makeScale( s, s, s );
-
-						}
-
-						if ( isInverse ) tempMatrix.invert();
-						matrix.multiply( tempMatrix );
-
-					}
-
-				} else if ( opName === 'xformOp:rotateXYZ' ) {
-
-					const r = data[ 'xformOp:rotateXYZ' ];
-					if ( r ) {
-
-						// USD rotateXYZ: matrix = Rx * Ry * Rz
-						// Three.js Euler 'ZYX' order produces Rx * Ry * Rz
-						const euler = new Euler(
-							r[ 0 ] * Math.PI / 180,
-							r[ 1 ] * Math.PI / 180,
-							r[ 2 ] * Math.PI / 180,
-							'ZYX'
-						);
-						tempMatrix.makeRotationFromEuler( euler );
-						if ( isInverse ) tempMatrix.invert();
-						matrix.multiply( tempMatrix );
-
-					}
-
-				} else if ( opName === 'xformOp:rotateX' ) {
-
-					const r = data[ 'xformOp:rotateX' ];
-					if ( r !== undefined ) {
-
-						tempMatrix.makeRotationX( r * Math.PI / 180 );
-						if ( isInverse ) tempMatrix.invert();
-						matrix.multiply( tempMatrix );
-
-					}
-
-				} else if ( opName === 'xformOp:rotateY' ) {
-
-					const r = data[ 'xformOp:rotateY' ];
-					if ( r !== undefined ) {
-
-						tempMatrix.makeRotationY( r * Math.PI / 180 );
-						if ( isInverse ) tempMatrix.invert();
-						matrix.multiply( tempMatrix );
-
-					}
-
-				} else if ( opName === 'xformOp:rotateZ' ) {
-
-					const r = data[ 'xformOp:rotateZ' ];
-					if ( r !== undefined ) {
-
-						tempMatrix.makeRotationZ( r * Math.PI / 180 );
-						if ( isInverse ) tempMatrix.invert();
-						matrix.multiply( tempMatrix );
-
-					}
-
-				}
-
-			}
-
-			obj.matrix.copy( matrix );
-			obj.matrix.decompose( obj.position, obj.quaternion, obj.scale );
-			return;
-
-		}
-
-		// Fallback: handle individual transform ops without order
-		if ( data[ 'xformOp:translate' ] ) {
-
-			const t = data[ 'xformOp:translate' ];
-			obj.position.set( t[ 0 ], t[ 1 ], t[ 2 ] );
-
-		}
-
-		if ( data[ 'xformOp:scale' ] ) {
-
-			const s = data[ 'xformOp:scale' ];
-
-			if ( Array.isArray( s ) ) {
-
-				obj.scale.set( s[ 0 ], s[ 1 ], s[ 2 ] );
-
-			} else {
-
-				obj.scale.set( s, s, s );
-
-			}
-
-		}
-
-		if ( data[ 'xformOp:rotateXYZ' ] ) {
-
-			const r = data[ 'xformOp:rotateXYZ' ];
-			obj.rotation.set(
-				r[ 0 ] * Math.PI / 180,
-				r[ 1 ] * Math.PI / 180,
-				r[ 2 ] * Math.PI / 180
-			);
-
-		}
-
-	}
-
-	_buildGeometry( path, fields, hasSkinning = false ) {
-
-		const geometry = new BufferGeometry();
-
-		const points = fields[ 'points' ];
-		if ( ! points || points.length === 0 ) return geometry;
-
-		const faceVertexIndices = fields[ 'faceVertexIndices' ];
-		const faceVertexCounts = fields[ 'faceVertexCounts' ];
-
-		let indices = faceVertexIndices;
-		if ( faceVertexCounts && faceVertexCounts.length > 0 ) {
-
-			indices = this._triangulateIndices( faceVertexIndices, faceVertexCounts );
-
-		}
-
-		let positions = points;
-		if ( indices && indices.length > 0 ) {
-
-			positions = this._expandAttribute( points, indices, 3 );
-
-		}
-
-		geometry.setAttribute( 'position', new BufferAttribute( new Float32Array( positions ), 3 ) );
-
-		const normals = fields[ 'normals' ] || fields[ 'primvars:normals' ];
-		if ( normals && normals.length > 0 ) {
-
-			let normalData = normals;
-			if ( normals.length === points.length ) {
-
-				// Per-vertex normals
-				if ( indices && indices.length > 0 ) {
-
-					normalData = this._expandAttribute( normals, indices, 3 );
-
-				}
-
-			} else if ( indices ) {
-
-				// Per-face-vertex normals
-				const normalIndices = this._triangulateIndices(
-					Array.from( { length: normals.length / 3 }, ( _, i ) => i ),
-					faceVertexCounts
-				);
-				normalData = this._expandAttribute( normals, normalIndices, 3 );
-
-			}
-
-			geometry.setAttribute( 'normal', new BufferAttribute( new Float32Array( normalData ), 3 ) );
-
-		} else {
-
-			geometry.computeVertexNormals();
-
-		}
-
-		const { uvs, uvIndices } = this._findUVPrimvar( fields );
-
-		if ( uvs && uvs.length > 0 ) {
-
-			let uvData = uvs;
-
-			if ( uvIndices && uvIndices.length > 0 ) {
-
-				// Custom UV indices
-				const triangulatedUvIndices = this._triangulateIndices( uvIndices, faceVertexCounts );
-				uvData = this._expandAttribute( uvs, triangulatedUvIndices, 2 );
-
-			} else if ( indices && uvs.length / 2 === points.length / 3 ) {
-
-				// Per-vertex UVs
-				uvData = this._expandAttribute( uvs, indices, 2 );
-
-			}
-
-			geometry.setAttribute( 'uv', new BufferAttribute( new Float32Array( uvData ), 2 ) );
-
-		}
-
-		// Add skinning attributes
-		if ( hasSkinning ) {
-
-			const jointIndices = fields[ 'primvars:skel:jointIndices' ];
-			const jointWeights = fields[ 'primvars:skel:jointWeights' ];
-			const elementSize = fields[ 'primvars:skel:jointIndices:elementSize' ] || 4;
-
-			if ( jointIndices && jointWeights ) {
-
-				const numVertices = positions.length / 3; // After expansion
-
-				// Expand skinning attributes by the same indices used for positions
-				let skinIndexData, skinWeightData;
-
-				if ( indices && indices.length > 0 ) {
-
-					skinIndexData = this._expandAttribute( jointIndices, indices, elementSize );
-					skinWeightData = this._expandAttribute( jointWeights, indices, elementSize );
-
-				} else {
-
-					skinIndexData = jointIndices;
-					skinWeightData = jointWeights;
-
-				}
-
-				// Three.js expects exactly 4 influences per vertex
-				const skinIndices = new Uint16Array( numVertices * 4 );
-				const skinWeights = new Float32Array( numVertices * 4 );
-
-				for ( let i = 0; i < numVertices; i ++ ) {
-
-					for ( let j = 0; j < 4; j ++ ) {
-
-						if ( j < elementSize ) {
-
-							skinIndices[ i * 4 + j ] = skinIndexData[ i * elementSize + j ] || 0;
-							skinWeights[ i * 4 + j ] = skinWeightData[ i * elementSize + j ] || 0;
-
-						} else {
-
-							skinIndices[ i * 4 + j ] = 0;
-							skinWeights[ i * 4 + j ] = 0;
-
-						}
-
-					}
-
-				}
-
-				geometry.setAttribute( 'skinIndex', new BufferAttribute( skinIndices, 4 ) );
-				geometry.setAttribute( 'skinWeight', new BufferAttribute( skinWeights, 4 ) );
-
-			}
-
-		}
-
-		return geometry;
-
-	}
-
-	_triangulateIndices( indices, counts ) {
-
-		const triangulated = [];
-		let offset = 0;
-
-		for ( let i = 0; i < counts.length; i ++ ) {
-
-			const count = counts[ i ];
-
-			if ( count === 3 ) {
-
-				triangulated.push(
-					indices[ offset ],
-					indices[ offset + 1 ],
-					indices[ offset + 2 ]
-				);
-
-			} else if ( count === 4 ) {
-
-				// Quad to two triangles
-				triangulated.push(
-					indices[ offset ],
-					indices[ offset + 1 ],
-					indices[ offset + 2 ],
-					indices[ offset ],
-					indices[ offset + 2 ],
-					indices[ offset + 3 ]
-				);
-
-			} else if ( count > 4 ) {
-
-				// Fan triangulation for n-gons
-				for ( let j = 1; j < count - 1; j ++ ) {
-
-					triangulated.push(
-						indices[ offset ],
-						indices[ offset + j ],
-						indices[ offset + j + 1 ]
-					);
-
-				}
-
-			}
-
-			offset += count;
-
-		}
-
-		return triangulated;
-
-	}
-
-	_expandAttribute( data, indices, itemSize ) {
-
-		const expanded = new Float32Array( indices.length * itemSize );
-
-		for ( let i = 0; i < indices.length; i ++ ) {
-
-			const srcIndex = indices[ i ] * itemSize;
-			const dstIndex = i * itemSize;
-
-			for ( let j = 0; j < itemSize; j ++ ) {
-
-				expanded[ dstIndex + j ] = data[ srcIndex + j ];
-
-			}
-
-		}
-
-		return expanded;
-
-	}
-
-	_buildMaterial( meshPath, fields ) {
-
-		const material = new MeshPhysicalMaterial();
-
-		// Try to find material binding
-		let materialPath = null;
-
-		// Check for material binding in fields
-		let materialBinding = fields[ 'material:binding' ];
-
-		// Check for relationship spec on mesh directly
-		if ( ! materialBinding ) {
-
-			const bindingPath = meshPath + '.material:binding';
-			const bindingSpec = this.specsByPath[ bindingPath ];
-			if ( bindingSpec && bindingSpec.specType === SpecType.Relationship ) {
-
-				materialBinding = bindingSpec.fields.targetPaths || bindingSpec.fields.default;
-
-			}
-
-		}
-
-		if ( materialBinding ) {
-
-			materialPath = Array.isArray( materialBinding ) ? materialBinding[ 0 ] : materialBinding;
-
-		}
-
-		// If no direct binding, check for GeomSubset children with material bindings
-		if ( ! materialPath ) {
-
-			const materialPaths = [];
-			const prefix = meshPath + '/';
-
-			for ( const path in this.specsByPath ) {
-
-				// Look for material:binding specs under mesh path
-				if ( ! path.startsWith( prefix ) ) continue;
-				if ( ! path.endsWith( '.material:binding' ) ) continue;
-
-				const bindingSpec = this.specsByPath[ path ];
-				if ( ! bindingSpec ) continue;
-
-				const targetPaths = bindingSpec.fields.targetPaths;
-				if ( targetPaths && targetPaths.length > 0 ) {
-
-					materialPaths.push( targetPaths[ 0 ] );
-
-				}
-
-			}
-
-			// Pick a material that has textures if possible
-			if ( materialPaths.length > 0 ) {
-
-				materialPath = this._pickBestMaterial( materialPaths );
-
-			}
-
-		}
-
-		// Fallback: try to find material in Looks hierarchy
-		if ( ! materialPath ) {
-
-			// Get root of mesh hierarchy (e.g., /chair_swan from /chair_swan/RedChairFeet)
-			const meshParts = meshPath.split( '/' );
-			const rootPath = '/' + meshParts[ 1 ];
-
-			// Look for materials in /Root/Looks/ or /Root/Materials/
-			for ( const path in this.specsByPath ) {
-
-				const spec = this.specsByPath[ path ];
-				if ( spec.specType !== SpecType.Prim ) continue;
-				if ( spec.fields.typeName !== 'Material' ) continue;
-
-				// Check if this material is in the same hierarchy
-				if ( path.startsWith( rootPath + '/Looks/' ) ||
-					path.startsWith( rootPath + '/Materials/' ) ) {
-
-					materialPath = path;
-					break;
-
-				}
-
-			}
-
-		}
-
-		if ( materialPath ) {
-
-			this._applyMaterial( material, materialPath );
-
-		}
-
-		return material;
-
-	}
-
-	_pickBestMaterial( materialPaths ) {
-
-		// Prefer materials that have texture files
-		for ( const materialPath of materialPaths ) {
-
-			const prefix = materialPath + '/';
-
-			// Check if this material has any texture shaders
-			for ( const path in this.specsByPath ) {
-
-				if ( ! path.startsWith( prefix ) ) continue;
-
-				const spec = this.specsByPath[ path ];
-				if ( spec.fields.typeName !== 'Shader' ) continue;
-
-				// Check for UsdUVTexture shader
-				const attrs = this._getAttributeValues( path );
-				if ( attrs[ 'info:id' ] === 'UsdUVTexture' && attrs[ 'inputs:file' ] ) {
-
-					return materialPath;
-
-				}
-
-			}
-
-		}
-
-		// Fallback to first material
-		return materialPaths[ 0 ];
-
-	}
-
-	_applyMaterial( material, materialPath ) {
-
-		const materialSpec = this.specsByPath[ materialPath ];
-		if ( ! materialSpec ) return;
-
-		const prefix = materialPath + '/';
-
-		// Look for shader children (UsdPreviewSurface)
-		for ( const path in this.specsByPath ) {
-
-			if ( ! path.startsWith( prefix ) ) continue;
-
-			const spec = this.specsByPath[ path ];
-			const typeName = spec.fields.typeName;
-
-			if ( typeName !== 'Shader' ) continue;
-
-			// Get shader attributes (info:id, inputs:*, etc.)
-			const shaderAttrs = this._getAttributeValues( path );
-
-			// Check for UsdPreviewSurface shader
-			const infoId = shaderAttrs[ 'info:id' ] || spec.fields[ 'info:id' ];
-
-			if ( infoId === 'UsdPreviewSurface' ) {
-
-				this._applyPreviewSurface( material, path );
-
-			}
-
-		}
-
-	}
-
-	_applyPreviewSurface( material, shaderPath ) {
-
-		const fields = this._getAttributeValues( shaderPath );
-
-		// Helper to get attribute spec with connection info
-		const getAttrSpec = ( attrName ) => {
-
-			const attrPath = shaderPath + '.' + attrName;
-			return this.specsByPath[ attrPath ];
-
-		};
-
-		// Helper to apply texture from connection
-		const applyTextureFromConnection = ( attrName, textureProperty, colorSpace, valueCallback ) => {
-
-			const spec = getAttrSpec( attrName );
-
-			if ( spec && spec.fields.connectionPaths && spec.fields.connectionPaths.length > 0 ) {
-
-				// Follow connection to texture shader
-				const connPath = spec.fields.connectionPaths[ 0 ];
-				const texture = this._getTextureFromConnection( connPath );
-
-				if ( texture ) {
-
-					texture.colorSpace = colorSpace;
-					material[ textureProperty ] = texture;
-					return true;
-
-				}
-
-			}
-
-			// No texture connection, use default value if present
-			if ( fields[ attrName ] !== undefined && valueCallback ) {
-
-				valueCallback( fields[ attrName ] );
-
-			}
-
-			return false;
-
-		};
-
-		// Diffuse color / base color map
-		applyTextureFromConnection(
-			'inputs:diffuseColor',
-			'map',
-			SRGBColorSpace,
-			( color ) => {
-
-				if ( Array.isArray( color ) && color.length >= 3 ) {
-
-					material.color.setRGB( color[ 0 ], color[ 1 ], color[ 2 ] );
-
-				}
-
-			}
-		);
-
-		// Emissive
-		applyTextureFromConnection(
-			'inputs:emissiveColor',
-			'emissiveMap',
-			SRGBColorSpace,
-			( color ) => {
-
-				if ( Array.isArray( color ) && color.length >= 3 ) {
-
-					material.emissive.setRGB( color[ 0 ], color[ 1 ], color[ 2 ] );
-
-				}
-
-			}
-		);
-
-		if ( material.emissiveMap ) {
-
-			material.emissive.set( 0xffffff );
-
-		}
-
-		// Normal map
-		applyTextureFromConnection( 'inputs:normal', 'normalMap', NoColorSpace, null );
-
-		// Roughness
-		const hasRoughnessMap = applyTextureFromConnection(
-			'inputs:roughness',
-			'roughnessMap',
-			NoColorSpace,
-			( value ) => {
-
-				material.roughness = value;
-
-			}
-		);
-
-		if ( hasRoughnessMap ) {
-
-			material.roughness = 1.0;
-
-		}
-
-		// Metallic
-		const hasMetalnessMap = applyTextureFromConnection(
-			'inputs:metallic',
-			'metalnessMap',
-			NoColorSpace,
-			( value ) => {
-
-				material.metalness = value;
-
-			}
-		);
-
-		if ( hasMetalnessMap ) {
-
-			material.metalness = 1.0;
-
-		}
-
-		// Occlusion
-		applyTextureFromConnection( 'inputs:occlusion', 'aoMap', NoColorSpace, null );
-
-		// IOR
-		if ( fields[ 'inputs:ior' ] !== undefined ) {
-
-			material.ior = fields[ 'inputs:ior' ];
-
-		}
-
-		// Clearcoat
-		if ( fields[ 'inputs:clearcoat' ] !== undefined ) {
-
-			material.clearcoat = fields[ 'inputs:clearcoat' ];
-
-		}
-
-		// Clearcoat roughness
-		if ( fields[ 'inputs:clearcoatRoughness' ] !== undefined ) {
-
-			material.clearcoatRoughness = fields[ 'inputs:clearcoatRoughness' ];
-
-		}
-
-		// Opacity / transparency
-		const opacitySpec = getAttrSpec( 'inputs:opacity' );
-
-		if ( opacitySpec && opacitySpec.fields.connectionPaths && opacitySpec.fields.connectionPaths.length > 0 ) {
-
-			const opacityConn = opacitySpec.fields.connectionPaths[ 0 ];
-
-			// Check if opacity is connected to alpha channel of diffuse texture
-			if ( opacityConn.endsWith( '.outputs:a' ) ) {
-
-				// Alpha is in the diffuse texture - enable transparency with blending
-				material.transparent = true;
-
-			} else {
-
-				// Separate opacity texture
-				const texture = this._getTextureFromConnection( opacityConn );
-				if ( texture ) {
-
-					texture.colorSpace = NoColorSpace;
-					material.alphaMap = texture;
-					material.transparent = true;
-
-				}
-
-			}
-
-		} else if ( fields[ 'inputs:opacity' ] !== undefined ) {
-
-			const opacity = fields[ 'inputs:opacity' ];
-			if ( typeof opacity === 'number' && opacity < 1.0 ) {
-
-				material.opacity = opacity;
-				material.transparent = true;
-
-			}
-
-		}
-
-	}
-
-	_getTextureFromConnection( connectionPath ) {
-
-		// connectionPath is like "/Material/TextureShader.outputs:rgb"
-		// Extract the shader path
-		const dotIdx = connectionPath.lastIndexOf( '.' );
-		if ( dotIdx === - 1 ) return null;
-
-		const textureShaderPath = connectionPath.slice( 0, dotIdx );
-		const textureShaderSpec = this.specsByPath[ textureShaderPath ];
-
-		if ( ! textureShaderSpec || textureShaderSpec.fields.typeName !== 'Shader' ) return null;
-
-		const textureAttrs = this._getAttributeValues( textureShaderPath );
-		const infoId = textureAttrs[ 'info:id' ];
-
-		if ( infoId !== 'UsdUVTexture' ) return null;
-
-		const file = textureAttrs[ 'inputs:file' ];
-		if ( ! file ) return null;
-
-		const texture = this._loadTexture( file );
-		if ( ! texture ) return null;
-
-		// Apply wrap modes
-		const wrapS = textureAttrs[ 'inputs:wrapS' ];
-		const wrapT = textureAttrs[ 'inputs:wrapT' ];
-
-		if ( wrapS ) texture.wrapS = this._getWrapMode( wrapS );
-		if ( wrapT ) texture.wrapT = this._getWrapMode( wrapT );
-
-		return texture;
-
-	}
-
-	_loadTexture( filePath ) {
-
-		// Clean up path
-		let cleanPath = filePath;
-		if ( cleanPath.startsWith( '@' ) ) cleanPath = cleanPath.slice( 1 );
-		if ( cleanPath.endsWith( '@' ) ) cleanPath = cleanPath.slice( 0, - 1 );
-		if ( cleanPath.startsWith( './' ) ) cleanPath = cleanPath.slice( 2 );
-
-		// Check cache first
-		if ( this.textureCache[ cleanPath ] ) {
-
-			return this.textureCache[ cleanPath ];
-
-		}
-
-		// Load from assets
-		const assetUrl = this.assets[ cleanPath ];
-		if ( assetUrl ) {
-
-			const texture = this.textureLoader.load( assetUrl );
-			this.textureCache[ cleanPath ] = texture;
-			return texture;
-
-		}
-
-		return null;
-
-	}
-
-	_getWrapMode( mode ) {
-
-		switch ( mode ) {
-
-			case 'clamp': return ClampToEdgeWrapping;
-			case 'mirror': return MirroredRepeatWrapping;
-			case 'repeat': return RepeatWrapping;
-			default: return RepeatWrapping;
-
-		}
-
-	}
-
-	// ========================================================================
-	// Skeletal Animation
-	// ========================================================================
-
-	_buildSkeleton( path ) {
-
-		const attrs = this._getAttributeValues( path );
-
-		// Get joint names (paths like "root", "root/body_joint", etc.)
-		const joints = attrs[ 'joints' ];
-		if ( ! joints || joints.length === 0 ) return null;
-
-		// Get bind transforms (world-space bind pose matrices)
-		const bindTransforms = attrs[ 'bindTransforms' ];
-		const restTransforms = attrs[ 'restTransforms' ];
-
-		// Build bones
-		const bones = [];
-		const bonesByPath = {};
-		const boneInverses = [];
-
-		for ( let i = 0; i < joints.length; i ++ ) {
-
-			const jointPath = joints[ i ];
-			const jointName = jointPath.split( '/' ).pop();
-
-			const bone = new Bone();
-			bone.name = jointName;
-			bones.push( bone );
-			bonesByPath[ jointPath ] = { bone, index: i };
-
-			// Compute inverse bind matrix
-			if ( bindTransforms && bindTransforms.length >= ( i + 1 ) * 16 ) {
-
-				const bindMatrix = new Matrix4();
-				// USD matrices are row-major. When loaded via fromArray into Three.js
-				// column-major storage, they get automatically transposed.
-				bindMatrix.fromArray( bindTransforms, i * 16 );
-				const inverseBindMatrix = bindMatrix.clone().invert();
-				boneInverses.push( inverseBindMatrix );
-
-			} else {
-
-				boneInverses.push( new Matrix4() );
-
-			}
-
-		}
-
-		// Build parent-child relationships based on joint paths
-		for ( let i = 0; i < joints.length; i ++ ) {
-
-			const jointPath = joints[ i ];
-			const parts = jointPath.split( '/' );
-
-			if ( parts.length > 1 ) {
-
-				const parentPath = parts.slice( 0, - 1 ).join( '/' );
-				const parentData = bonesByPath[ parentPath ];
-
-				if ( parentData ) {
-
-					parentData.bone.add( bones[ i ] );
-
-				}
-
-			}
-
-		}
-
-		// Apply rest transforms to bones (local transforms)
-		if ( restTransforms && restTransforms.length >= joints.length * 16 ) {
-
-			for ( let i = 0; i < joints.length; i ++ ) {
-
-				const matrix = new Matrix4();
-				// USD matrices are row-major. When loaded via fromArray into Three.js
-				// column-major storage, they get automatically transposed.
-				matrix.fromArray( restTransforms, i * 16 );
-				matrix.decompose( bones[ i ].position, bones[ i ].quaternion, bones[ i ].scale );
-
-			}
-
-		}
-
-		// Find root bone(s) - bones without a parent bone
-		const rootBones = bones.filter( bone => ! bone.parent || ! bone.parent.isBone );
-
-		// Get animation source path
-		const animSourceSpec = this.specsByPath[ path + '.skel:animationSource' ];
-		let animationPath = null;
-		if ( animSourceSpec && animSourceSpec.fields.targetPaths && animSourceSpec.fields.targetPaths.length > 0 ) {
-
-			animationPath = animSourceSpec.fields.targetPaths[ 0 ];
-
-		}
-
-		return {
-			skeleton: new Skeleton( bones, boneInverses ),
-			joints: joints,
-			rootBones: rootBones,
-			animationPath: animationPath,
-			path: path
-		};
-
-	}
-
-	_bindSkeletons() {
-
-		for ( const meshData of this.skinnedMeshes ) {
-
-			const { mesh, skeletonPath, localJoints } = meshData;
-
-			let skeletonData = null;
-			for ( const skelPath in this.skeletons ) {
-
-				if ( skeletonPath && skeletonPath.includes( skelPath ) ) {
-
-					skeletonData = this.skeletons[ skelPath ];
-					break;
-
-				}
-
-			}
-
-			// Fallback to first skeleton for single-skeleton files
-			if ( ! skeletonData ) {
-
-				const skeletonPaths = Object.keys( this.skeletons );
-				if ( skeletonPaths.length > 0 ) {
-
-					skeletonData = this.skeletons[ skeletonPaths[ 0 ] ];
-
-				}
-
-			}
-
-			if ( skeletonData ) {
-
-				const { skeleton, rootBones, joints } = skeletonData;
-
-				// Remap local joint indices to global skeleton indices
-				// Each mesh can have its own skel:joints array that defines which
-				// subset of skeleton joints it uses (and in what order)
-				if ( localJoints && localJoints.length > 0 ) {
-
-					const skinIndex = mesh.geometry.attributes.skinIndex;
-					if ( skinIndex ) {
-
-						// Build mapping: local index -> global skeleton index
-						const localToGlobal = [];
-						for ( let i = 0; i < localJoints.length; i ++ ) {
-
-							const jointName = localJoints[ i ];
-							const globalIdx = joints.indexOf( jointName );
-							localToGlobal[ i ] = globalIdx >= 0 ? globalIdx : 0;
-
-						}
-
-						// Remap all joint indices
-						const arr = skinIndex.array;
-						for ( let i = 0; i < arr.length; i ++ ) {
-
-							const localIdx = arr[ i ];
-							if ( localIdx < localToGlobal.length ) {
-
-								arr[ i ] = localToGlobal[ localIdx ];
-
-							}
-
-						}
-
-					}
-
-				}
-
-				// Add root bones to the mesh first
-				for ( const rootBone of rootBones ) {
-
-					mesh.add( rootBone );
-
-				}
-
-				// Bind the skeleton to the mesh with identity bind matrix
-				// We pass a bind matrix to prevent Three.js from overwriting
-				// our carefully computed boneInverses via calculateInverses()
-				mesh.bind( skeleton, new Matrix4() );
-
-			}
-
-		}
-
-	}
-
-	_buildAnimations() {
-
-		const animations = [];
-
-		// Find all SkelAnimation prims
-		for ( const path in this.specsByPath ) {
-
-			const spec = this.specsByPath[ path ];
-			if ( spec.specType !== SpecType.Prim ) continue;
-			if ( spec.fields.typeName !== 'SkelAnimation' ) continue;
-
-			const clip = this._buildAnimationClip( path );
-			if ( clip ) {
-
-				animations.push( clip );
-
-			}
-
-		}
-
-		return animations;
-
-	}
-
-	_buildAnimationClip( path ) {
-
-		const attrs = this._getAttributeValues( path );
-		const joints = attrs[ 'joints' ];
-
-		if ( ! joints || joints.length === 0 ) return null;
-
-		const tracks = [];
-
-		// Get rotation time samples
-		const rotationsAttr = this._getTimeSampledAttribute( path, 'rotations' );
-		if ( rotationsAttr && rotationsAttr.times && rotationsAttr.values ) {
-
-			const { times, values } = rotationsAttr;
-
-			for ( let jointIdx = 0; jointIdx < joints.length; jointIdx ++ ) {
-
-				const jointName = joints[ jointIdx ].split( '/' ).pop();
-				const keyframeTimes = [];
-				const keyframeValues = [];
-
-				for ( let t = 0; t < times.length; t ++ ) {
-
-					const quatData = values[ t ];
-					if ( ! quatData || quatData.length < ( jointIdx + 1 ) * 4 ) continue;
-
-					keyframeTimes.push( times[ t ] / this.fps );
-
-					// USD GfQuatf stores imaginary (x,y,z) first, then real (w)
-					// This matches Three.js quaternion order (x,y,z,w)
-					const x = quatData[ jointIdx * 4 + 0 ];
-					const y = quatData[ jointIdx * 4 + 1 ];
-					const z = quatData[ jointIdx * 4 + 2 ];
-					const w = quatData[ jointIdx * 4 + 3 ];
-					keyframeValues.push( x, y, z, w );
-
-				}
-
-				if ( keyframeTimes.length > 0 ) {
-
-					tracks.push( new QuaternionKeyframeTrack(
-						jointName + '.quaternion',
-						new Float32Array( keyframeTimes ),
-						new Float32Array( keyframeValues )
-					) );
-
-				}
-
-			}
-
-		}
-
-		// Get translation time samples
-		const translationsAttr = this._getTimeSampledAttribute( path, 'translations' );
-		if ( translationsAttr && translationsAttr.times && translationsAttr.values ) {
-
-			const { times, values } = translationsAttr;
-
-			for ( let jointIdx = 0; jointIdx < joints.length; jointIdx ++ ) {
-
-				const jointName = joints[ jointIdx ].split( '/' ).pop();
-				const keyframeTimes = [];
-				const keyframeValues = [];
-
-				for ( let t = 0; t < times.length; t ++ ) {
-
-					const transData = values[ t ];
-					if ( ! transData || transData.length < ( jointIdx + 1 ) * 3 ) continue;
-
-					keyframeTimes.push( times[ t ] / this.fps );
-					keyframeValues.push(
-						transData[ jointIdx * 3 + 0 ],
-						transData[ jointIdx * 3 + 1 ],
-						transData[ jointIdx * 3 + 2 ]
-					);
-
-				}
-
-				if ( keyframeTimes.length > 0 ) {
-
-					tracks.push( new VectorKeyframeTrack(
-						jointName + '.position',
-						new Float32Array( keyframeTimes ),
-						new Float32Array( keyframeValues )
-					) );
-
-				}
-
-			}
-
-		}
-
-		// Get scale time samples
-		const scalesAttr = this._getTimeSampledAttribute( path, 'scales' );
-		if ( scalesAttr && scalesAttr.times && scalesAttr.values ) {
-
-			const { times, values } = scalesAttr;
-
-			for ( let jointIdx = 0; jointIdx < joints.length; jointIdx ++ ) {
-
-				const jointName = joints[ jointIdx ].split( '/' ).pop();
-				const keyframeTimes = [];
-				const keyframeValues = [];
-
-				for ( let t = 0; t < times.length; t ++ ) {
-
-					const scaleData = values[ t ];
-					if ( ! scaleData || scaleData.length < ( jointIdx + 1 ) * 3 ) continue;
-
-					keyframeTimes.push( times[ t ] / this.fps );
-					keyframeValues.push(
-						scaleData[ jointIdx * 3 + 0 ],
-						scaleData[ jointIdx * 3 + 1 ],
-						scaleData[ jointIdx * 3 + 2 ]
-					);
-
-				}
-
-				if ( keyframeTimes.length > 0 ) {
-
-					tracks.push( new VectorKeyframeTrack(
-						jointName + '.scale',
-						new Float32Array( keyframeTimes ),
-						new Float32Array( keyframeValues )
-					) );
-
-				}
-
-			}
-
-		}
-
-		if ( tracks.length === 0 ) return null;
-
-		const clipName = this._getPathName( path );
-		return new AnimationClip( clipName, - 1, tracks );
-
-	}
-
-	_getTimeSampledAttribute( primPath, attrName ) {
-
-		// Look for the attribute spec with time samples
-		const attrPath = primPath + '.' + attrName;
-		const attrSpec = this.specsByPath[ attrPath ];
-
-		if ( attrSpec && attrSpec.fields.timeSamples ) {
-
-			const timeSamples = attrSpec.fields.timeSamples;
-			if ( timeSamples.times && timeSamples.values ) {
-
-				return timeSamples;
-
-			}
-
-		}
-
-		return null;
 
 	}
